# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/careers": {
            "get": {
                "description": "Obtiene todas las carreras de la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carreras"
                ],
                "summary": "Obtiene todas las carreras",
                "responses": {
                    "200": {
                        "description": "Carreras obtenidas exitosamente",
                        "schema": {
                            "$ref": "#/definitions/responses.CareerResponse"
                        }
                    },
                    "400": {
                        "description": "Error al procesar la solicitud",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/student": {
            "post": {
                "description": "Registra un nuevo estudiante en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Registra un nuevo estudiante",
                "parameters": [
                    {
                        "description": "Estudiante a registrar",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Estudiante"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estudiante creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/responses.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Error al procesar la solicitud",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/teacher": {
            "post": {
                "description": "Registra un nuevo profesor en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profesores"
                ],
                "summary": "Registra un nuevo profesor",
                "parameters": [
                    {
                        "description": "Profesor a registrar",
                        "name": "teacher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Profesor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profesor creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/responses.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Error al procesar la solicitud",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/zodiacal-signs": {
            "get": {
                "description": "Obtiene todos los signos zodiacales de la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Signos Zodiacales"
                ],
                "summary": "Obtiene todos los signos zodiacales",
                "responses": {
                    "200": {
                        "description": "Signos zodiacales obtenidos exitosamente",
                        "schema": {
                            "$ref": "#/definitions/responses.ZodiacalSignResponse"
                        }
                    },
                    "400": {
                        "description": "Error al procesar la solicitud",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Carrera": {
            "type": "object",
            "properties": {
                "director": {
                    "type": "string"
                },
                "duracion": {
                    "type": "integer"
                },
                "estudiantes_registrados": {
                    "type": "integer"
                },
                "facultad": {
                    "type": "string"
                },
                "nombre_carrera": {
                    "type": "string"
                }
            }
        },
...
        "models.Estudiante": {
            "type": "object",
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "carnet": {
                    "type": "string"
                },
                "colegio": {
                    "type": "string"
                },
                "conexiones": {
                    "description": "Conexiones usuarios",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "correo": {
                    "type": "string"
                },
                "fecha_nacimiento": {
                    "type": "string"
                },
                "foraneo": {
                    "type": "boolean"
                },
                "genero": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "parqueo": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "publicaciones": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "usuario": {
                    "type": "string"
                }
            }
        },
...
        "models.Profesor": {
            "type": "object",
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "conexiones": {
                    "description": "Conexiones usuarios",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "correo": {
                    "type": "string"
                },
                "departamento": {
                    "type": "string"
                },
                "fecha_nacimiento": {
                    "type": "string"
                },
                "genero": {
                    "type": "string"
                },
                "jornada": {
                    "type": "string"
                },
                "maestria": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "publicaciones": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "usuario": {
                    "type": "string"
                }
            }
        },
        "models.Signo": {
            "type": "object",
            "properties": {
                "dia_semana": {
                    "type": "string"
                },
                "elemento": {
                    "type": "string"
                },
                "metal": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "piedra": {
                    "type": "string"
                },
                "planeta": {
                    "type": "string"
                }
            }
        },
        "responses.CareerResponse": {
            "type": "object",
            "properties": {
                "careers": {
                    "description": "Datos adicionales de la respuesta",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Carrera"
                    }
                },
                "message": {
                    "description": "Mensaje de la respuesta",
                    "type": "string"
                },
                "status": {
                    "description": "Código de estado de la respuesta",
                    "type": "integer"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Detalles específicos del error",
                    "type": "string"
                },
                "message": {
                    "description": "Mensaje de error de la respuesta",
                    "type": "string"
                },
                "status": {
                    "description": "Código de error de la respuesta",
                    "type": "integer"
                }
            }
        },
...
        "responses.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Datos adicionales de la respuesta",
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "description": "Mensaje de la respuesta",
                    "type": "string"
                },
                "status": {
                    "description": "Código de estado de la respuesta",
                    "type": "integer"
                }
            }
        },
...
        "responses.ZodiacalSignResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensaje de la respuesta",
                    "type": "string"
                },
                "status": {
                    "description": "Código de estado de la respuesta",
                    "type": "integer"
                },
                "zodiacalSign": {
                    "description": "Datos adicionales de la respuesta",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Signo"
                    }
                }
            }
        }
    }
}